pyproject.toml
[tool.poetry]
name = "loader"
version = "0.1.0"
description = "The module for loading new datasets with appropriate logging, metadata tracking, and error handling."
authors = ["mikevatd3 <mike@datadrivendetroit.org>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
pandas = "^2.2.2"
python-dateutil = "^2.9.0"
geopandas = "^1.0.1"
tomli = "^2.0.1"
SQLAlchemy = "^2.0.32"
GeoAlchemy = "^0.7.2"
psycopg2 = "^2.9.9"
click = "^8.1.7"
returns = "^0.23.0"
metadata = {path = "/home/michael/projects/new/system/tools/metadata", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
